fix(ci): maintainability gate — split test_adapters.py under the 1500-line limit
A later Ollama change pushed tests/unit/adapters/test_adapters.py back to
1501 lines, re-tripping the first-party file line-limit guardrail. Move
TestCursorAdapter into test_cursor_adapter.py (mirroring the existing
test_grok_adapter.py split) and drop the now-unused model_selection import.
